es3???

+++ src/ES2/entities/CalcConsume.java
@@ -2,22 +2,34 @@ package ES2.entities;
 
 public class CalcConsume {
     //ATTRIBUTI
-    private double kmRun;
-    private double fuelConsume;
+    private int kmRun;
+    private int fuelConsume;
+//    private double kmRun;
+//    private double fuelConsume;
 
 
     //CISTRUTTORI
-    public CalcConsume(double kmRun, double fuelConsume) {
+    public CalcConsume(int kmRun, int fuelConsume) {
         this.kmRun = kmRun;
         this.fuelConsume = fuelConsume;
     }
+//    public CalcConsume(double kmRun, double fuelConsume) {
+//        this.kmRun = kmRun;
+//        this.fuelConsume = fuelConsume;
+//    }
 
     //METODI
-
-    public double kmPerLiter() throws ArithmeticException {
+    public int kmPerLiter() throws ArithmeticException {
         if (fuelConsume == 0) {
             throw new ArithmeticException("Complimenti, sei appena diventato l'uomo più ricco del mondo");
         }
         return kmRun / fuelConsume;
     }
+
+//    public double kmPerLiter() throws ArithmeticException {
+//        if (fuelConsume == 0) {
+//            throw new ArithmeticException("Complimenti, sei appena diventato l'uomo più ricco del mondo");
+//        }
+//        return kmRun / fuelConsume;
+//    }
 }
